# perplexity: avoid unnecessary alloocations and logit copies