This speeds up perplexity calculations by a large margin. E.g., for computing perplexity of `wiki.test.raw` for `fp16` Mistral-7B on RTX-4080 and 32-core Ryzen 5975WX CPU
* Time goes down from 70 seconds to 48 seconds for context of 512 (46% speedup)
* Time goes down from 114 seconds to 76 seconds for context of 4096 (49% speedup)