@Artefact2 2024-01-20
Considering imatrix shares a lot of code with perplexity, could a similar optimisation also apply to the former?